completed first draft
