+++ app/frontend/src/index.scss
@@ -14,5 +14,4 @@
 
 // Import page styles
 @use "pages/Demo/Demo";
-@use "pages/LandingPage/LandingPage";
-@use "pages/LandingPage/InnerCopNav";
+@use "pages/LandingPage/" as *;
