--- app/frontend/src/pages/LandingPage/_InnerCopNav.scss
@@ -1,5 +1,0 @@
-.inner-cop-nav {
-  display: flex;
-  flex-direction: column;
-  row-gap: 32px;
-}
