+++ app/frontend/src/pages/LandingPage/_index.scss
@@ -1,0 +1,2 @@
+@forward "LandingPageIntro";
+@forward "LandingPageCop";
